Split indicators app models into separate modules
Will ease problems with circular imports
